Update CupsOfCoffee.java
This update is the completed project.  It imports the DecimalFormat class to format numbers to "0.00", and the Scanner class to log input from the user.  It establishes a number of variables, asks the user for some information, builds a table to display their subtotal, tax, and total for the transaction that is decimal aligned.  This program applies table formatting, decimal formatting, print/ln/f statements, keyboard scanner, if statements, a switch statement, and an if-else-if set of statements.